refactor split add link button
